fix: change journalctl_header from parser to datasource
